requirements.txt
scapy~=2.6.1
pytest~=8.3.4
requests~=2.32.3
Flask~=3.1.0
PyYAML~=6.0.2